qr_code_flask_app/requirements.txt
rewuiFlask
flask_sqlalchemy
opencv-python
pyzbar